+++ src/settings/utils/tokens-reducer.js
@@ -1,7 +1,13 @@
 import { reduce } from 'lodash';
 
+const getSectionTokens = (section) => {
+  return reduce(section, (tokensList, { token, previewValue }) => {
+    return { ...tokensList, ...{ [token]: previewValue } };
+  }, {});
+};
+
 export default (tokens) => {
-  return reduce(tokens, (acc, category) => {
-    return { ...acc, ...category };
+  return reduce(tokens, (tokensList, section) => {
+    return { ...tokensList, ...getSectionTokens(section) };
   }, {});
 };